create alternative solution using inheritance
In this solution there will have been made changes to Item.
It is now a abstract base class. This is done to make code
more readable aswell as making it easier to implement new
items which do not have the exact name.

All items can also now be conjured items.